@@ -1,13 +1,13 @@
 module github.com/richardwilkes/gp
 
-go 1.25.0
+go 1.26.0
 
 require (
-	github.com/richardwilkes/toolbox/v2 v2.5.0
+	github.com/richardwilkes/toolbox/v2 v2.9.0
 	github.com/yookoala/realpath v1.0.0
 )
 
 require (
-	golang.org/x/sys v0.35.0 // indirect
-	golang.org/x/term v0.34.0 // indirect
+	golang.org/x/sys v0.41.0 // indirect
+	golang.org/x/term v0.40.0 // indirect
 )
